requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.50
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.10.0
future==1.0.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-aiplatform==1.126.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-cloud-vision==3.11.0
google-crc32c==1.7.1
google-genai==1.49.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
modelscope==1.31.0
networkx==3.5
numpy==2.3.4
openai==2.7.2
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.1
paddlepaddle==3.2.1
paddlex==3.3.9
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pypdfium2==5.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
gunicorn
